+++ FinancialTracker.Services.Analytics/Messaging/ITransactionCreatedMessageExtension.cs
@@ -3,9 +3,9 @@ using MessageBus.Shared.Contracts.Interfaces;
 
 namespace FinancialTracker.Services.Analytics.Messaging;
 
-public static class BusMessageExtensions
+public static class IExpenseCreatedMessageExtension
 {
-    public static string GetInfo(this ITransactionCreatedMessage message)
+    public static string GetInfo(this IExpenseCreatedMessage message)
     {
         var info = "";
         var messageType = message.GetType();
@@ -16,15 +16,15 @@ public static class BusMessageExtensions
 
         return info;
     }
-    public static ExpenseDto ToExpenseDto(this ITransactionCreatedMessage message)
+    public static ExpenseDto ToExpenseDto(this IExpenseCreatedMessage message)
     {
         return new ExpenseDto()
         {
             UserId = message.UserId,
             AccountId = message.AccountId,
-            ExpenseId = message.TransactionId,
+            ExpenseId = message.ExpenseId,
             Category = message.CategoryName,
-            ExpenseTime = message.TransactionDate,
+            ExpenseTime = message.ExpenseTime,
             Amount = message.Amount
         };
     }
